+++ app/src/main/kotlin/com/pitchedapps/frost/activities/ImagePickerActivity.kt
@@ -3,15 +3,10 @@ package com.pitchedapps.frost.activities
 import android.content.res.ColorStateList
 import android.os.Bundle
 import ca.allanwang.kau.imagepicker.ImagePickerActivityBase
+import ca.allanwang.kau.imagepicker.ImagePickerActivityOverlayBase
 import com.pitchedapps.frost.utils.Prefs
 
 /**
  * Created by Allan Wang on 2017-07-23.
  */
-class ImagePickerActivity : ImagePickerActivityBase() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        fab.backgroundTintList = ColorStateList.valueOf(Prefs.iconBackgroundColor)
-    }
-}
+class ImagePickerActivity : ImagePickerActivityOverlayBase()